test: move benchmark.js from common/ folder
remove benchmark.js from test/common to test/benchmark
change import part in all test-benchmark-* file
