+++ test/benchmark/test-benchmark-assert.js
@@ -5,6 +5,6 @@ require('../common');
 // Minimal test for assert benchmarks. This makes sure the benchmarks aren't
 // completely broken but nothing more than that.
 
-const runBenchmark = require('../common/benchmark');
+const runBenchmark = require('./benchmark');
 
 runBenchmark('assert');
