+++ test/benchmark/test-benchmark-dgram.js
@@ -2,7 +2,7 @@
 
 require('../common');
 
-const runBenchmark = require('../common/benchmark');
+const runBenchmark = require('./benchmark');
 
 // Dgram benchmarks use hardcoded ports. Thus, this test can not be run in
 // parallel with tests that choose random ports.
