+++ test/benchmark/test-benchmark-net.js
@@ -6,6 +6,6 @@ require('../common');
 // rather than parallel to make sure it does not conflict with tests that choose
 // random available ports.
 
-const runBenchmark = require('../common/benchmark');
+const runBenchmark = require('./benchmark');
 
 runBenchmark('net', { NODEJS_BENCHMARK_ZERO_ALLOWED: 1 });
