+++ test/benchmark/test-benchmark-policy.js
@@ -2,7 +2,7 @@
 
 require('../common');
 
-const runBenchmark = require('../common/benchmark');
+const runBenchmark = require('./benchmark');
 
 runBenchmark('policy', [
   'n=1',
