+++ test/benchmark/test-benchmark-v8.js
@@ -2,6 +2,6 @@
 
 require('../common');
 
-const runBenchmark = require('../common/benchmark');
+const runBenchmark = require('./benchmark');
 
 runBenchmark('v8', { NODEJS_BENCHMARK_ZERO_ALLOWED: 1 });
